Update SpurGear.fs to v2.3.0
Improved Addendum, Dedendum, and Chamfer options.

Added Custom option to Addendum and Dedendum Factor options that let you
specify the factor numerically and removed the now redundant tip and root
radius settings. Added a "Factor and angle" option to the Chamfer options that
lets you specify the chamfer distance as a multiple of the module so it scales
with the gear size.

Slightly improved/simplified the conversion of addendum/dedendum/chamfer
options to numeric values by using maps.

Changed the "Spur Gears" configuration to be just "Ring Gear" and made it more
configurable for non-standard and more efficient gears.